backend/requirements.txt
# FastAPI and ASGI server (upgraded for Google ADK compatibility)
fastapi>=0.115.0
uvicorn[standard]>=0.27.0

# CORS and basic middleware
python-multipart>=0.0.6

# Environment variables
python-dotenv>=1.0.1

# Testing (optional for now)
httpx>=0.26.0

# Google Agent Development Kit (ADK) for multi-agent system
google-adk==1.0.0

# ADK missing dependencies (should be auto-installed but aren't)
deprecated>=1.2.14

# WebSocket support for real-time debate streaming
websockets>=14.1

# Google Cloud Text-to-Speech for AI voices
google-cloud-texttospeech>=2.14.0

# Google Cloud Speech-to-Text for voice input
google-cloud-speech>=2.20.0

# LangChain for RAG (Retrieval Augmented Generation)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0

# Vector store and embeddings for RAG
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Wikipedia loader for custom figures
wikipedia>=1.4.0